Add toggle wire frame operator
